+++ releases.md
@@ -1,3 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - `String::Format.statistics` takes pairs argument instead of keyword argument.
+
 ## v0.1.0
+
+  - Initial release.